\name{testdata}
\alias{testdata}
\docType{data}
\title{
a data to test 
}
\description{
a data to test, X is 95x1000 genotype matrix, Y is a list of 4 phenotypes of 95 samples 
}
\usage{data("testdata")}
\format{
    X is 95x1000 genotype matrix, Y is a list of 4 phenotypes of 95 samples 
}
\examples{
data(testdata)
## maybe str(X) 
}
\keyword{datasets}
